scripts/scraper/requirements.txt
crawl4ai>=0.4.0
openai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
